Snow compiler is now a bit more strict with conflicting definitions/implementations. Fix the warning for long long vs NSUIntger

+++ core/S3AddBucketOperation.m
@@ -85,7 +85,7 @@ static NSString *S3LocationFormatString = @"<CreateBucketConfiguration><Location
     return nil;
 }
 
-- (long long)requestBodyContentLength
+- (NSUInteger)requestBodyContentLength
 {
     NSData *contents = [self requestBodyContentData];
     if (!contents) {
